manage_app/api/serializers/user_serializer.py
from rest_framework import serializers

from manage_app.models import CustomUser


class UserSerializer(serializers.ModelSerializer):
    class Meta:
        model = CustomUser
        fields = ('username', 'email', 'coins')
