requirements.txt
asgiref>=3.8.1
attrs>=25.3.0
Django==5.1.4
djangorestframework>=3.16.0
python-dotenv>=1.1.0
drf-spectacular>=0.28.0
inflection>=0.5.1
jsonschema>=4.23.0
jsonschema-specifications>=2024.10.1
PyYAML>=6.0.2
referencing>=0.36.2
rpds-py>=0.24.0
sqlparse>=0.5.3
typing_extensions>=4.13.2
tzdata>=2025.2
uritemplate>=4.1.1
django-cors-headers>=4.7.0
djangorestframework_simplejwt>=5.5.0
redis>=5.0.0
celery>=5.4.0
django-celery-beat>=2.7.0
django-celery-results>=2.5.1
psycopg2-binary>=2.9.9
gunicorn>=21.2.0
whitenoise==6.9.0
